Fix icons in HTTPS
The headers file contained a request to a http link which is bad if we are to use HTTPS. This resulted in the icons not loading. The link was changed to use HTTPS as we can always hit HTTPS from HTTP but not the other way round. Moreover, some issues in the python code was fixed to handle floats and convert it into an int